Working build, used a modal to cover spa. Lots of extra stuff to get rid of
