+++ src/index.js
@@ -7,9 +7,9 @@ import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
   <React.StrictMode>
-    <Router basename={process.env.PUBLIC_URL}>
+    
       <App/>
-    </Router>
+    
   </React.StrictMode>,
   document.getElementById('root')
 );
